# go/model_extension.go
/*
 * Extension Config API
 *
 * API to submit and publish an ExtensionConfig object
 *
 * API version: 1.0.0
 * Generated by: OpenAPI Generator (https://openapi-generator.tech)
 */

package openapi

type Extension struct {

	IsDownloaded bool `json:"isDownloaded,omitempty"`

	UserName string `json:"userName"`

	UserId string `json:"userId"`

	HasIcon bool `json:"hasIcon,omitempty"`

	// this is a unique name of the package. On extension config, it is set to kebab-case of user's set name
	Name string `json:"name"`

	Config string `json:"config"`

	PackageJson string `json:"packageJson"`

	Id float32 `json:"id,omitempty"`

	IsPublished bool `json:"isPublished,omitempty"`
}